Moved `EventStoreManagedUnitOfWork` to a separate source file
